requirements.txt
hdx-python-scraper[pandas]==1.7.3
